Add files via upload
in this project, first of all, understand the problem, then collect the data and clean it.
learn about the top 10 venues in which the same team won the match by winning the toss.
in all the matches that have gone to be played so far, who has won bat first and field first?
which is the team that has won the most matches, but its difference is how many matches it has won by batting first and how many matches it has won by field first, I got to learn this.
so far, I have learned to remove the top 10 list in so many seasons.
after that, find out who has won the top 10 MOM Winning Awards and how many times.
in this project, I learned about pivot tables and pivot charts and how to use them. 
learn how to create a dashboard, learn about the calculation field, create KPI (key point indicator), and learn about top 10 analysis.